utils.c: Use long rather than int for random() return value
This matches the random(3) definition even though with the % 64, we
only get a number in the range 0-63 back.

+++ www/utils.c
@@ -676,7 +676,7 @@ char *generate_password_hash(int hash_type, const char *password)
 		strcpy(salt, "$6$");
 
 	for (i = 3; i < 19; i++) {
-		int r;
+		long r;
 		struct timeval tv;
 
 		gettimeofday(&tv, NULL);
